refactor: grid constants naming
