+++ src/lib/tabs/tabOptimizer/optimizerForm/components/CombatBuffsDrawer.tsx
@@ -2,7 +2,7 @@ import { Drawer, Flex, Form, Typography } from 'antd'
 import { CombatBuffs } from 'lib/constants/constants'
 import { defaultGap } from 'lib/constants/constantsUi'
 import InputNumberStyled from 'lib/tabs/tabOptimizer/optimizerForm/components/InputNumberStyled'
-import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridConstants'
+import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridColumns'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
