+++ src/lib/tabs/tabOptimizer/optimizerForm/components/ComboFilter.tsx
@@ -3,7 +3,7 @@ import { Button, Flex, Form, Input, Popconfirm, Radio, Segmented } from 'antd'
 import { FormInstance } from 'antd/es/form/hooks/useForm'
 import DB from 'lib/state/db'
 import InputNumberStyled from 'lib/tabs/tabOptimizer/optimizerForm/components/InputNumberStyled'
-import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridConstants'
+import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridColumns'
 import { ComboDrawer } from 'lib/tabs/tabOptimizer/rotation/ComboDrawer'
 import { HeaderText } from 'lib/ui/HeaderText'
 import { useMemo } from 'react'
